requirements.txt
altair==4.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
bokeh==2.1.1
boto3==1.14.30
botocore==1.17.30
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
mlxtend==0.17.3
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numpy==1.19.1
oauthlib==3.1.0
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pathlib==1.0.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyarrow==1.0.0
pydeck==0.4.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
spacy==2.3.2
streamlit==0.64.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
tweepy==3.9.0
tzlocal==2.1
urllib3==1.25.10
validators==0.16.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.7.0
zipp==3.1.0
